+++ src/main/java/lotto/view/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view.domain;
 
 import java.util.List;
